task_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.0.0
python-jose[cryptography]==3.3.0
python-multipart
pika==1.3.2